backend/requirements.txt
aiofiles==25.1.0
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
fastapi-cache2==0.2.2
fastapi-mail==1.6.2
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pendulum==3.2.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2025.11.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
